converted components to jsx
