+++ src/components/App.jsx
@@ -1,8 +1,10 @@
 import React, {PureComponent} from 'react';
 import Page from './Page';
 
-export default class App extends PureComponent {
+export default class App extends React.PureComponent {
     render() {
+      return(
         return (<Page />);
+      );
     }
 };
